+++ src/Datasource/Paging/Exception/PageOutOfBoundsException.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
  * @since         3.5.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Datasource\Exception;
+namespace Cake\Datasource\Paging\Exception;
 
 use Cake\Core\Exception\CakeException;
