+++ src/Datasource/Paging/PaginatedInterface.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
  * @since         5.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Paging;
+namespace Cake\Datasource\Paging;
 
 use Countable;
 use Traversable;
