+++ src/Datasource/Paging/PaginatedResultSet.php
@@ -14,9 +14,9 @@ declare(strict_types=1);
  * @since         5.0.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Datasource;
+namespace Cake\Datasource\Paging;
 
-use Cake\Paging\PaginatedInterface;
+use Cake\Datasource\ResultSetInterface;
 use IteratorIterator;
 
 /**
