+++ tests/test_app/TestApp/Datasource/CustomPaginator.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace TestApp\Datasource;
 
-use Cake\Datasource\Paginator;
+use Cake\Datasource\Paging\Paginator;
 
 class CustomPaginator extends Paginator
 {
